+++ lib-artnet/include/artnetstore.h
@@ -35,35 +35,27 @@
 #include "configstore.h"
 #include "firmware/debug/debug_debug.h"
 
-namespace artnet::store
-{
-namespace internal
-{
-inline void SaveUniverse(uint32_t port_index)
-{
+namespace artnet::store {
+namespace internal {
+inline void SaveUniverse(uint32_t port_index) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
     uint16_t universe;
 
-    if (ArtNetNode::Get()->GetPortAddress(port_index, universe))
-    {
+    if (ArtNetNode::Get()->GetPortAddress(port_index, universe)) {
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-        if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-        {
+        if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
             port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-        }
-        else
-        {
+        } else {
             DEBUG_EXIT();
             return;
         }
 #endif
 
-        DEBUG_PRINTF("port_index=%u", port_index);
+        DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-        if (port_index >= artnet::kPorts)
-        {
+        if (port_index >= artnet::kPorts) {
             DEBUG_EXIT();
             return;
         }
@@ -73,7 +65,7 @@ inline void SaveUniverse(uint32_t port_index)
 #pragma GCC diagnostic ignored "-Wdangling-pointer"
 #endif
 
-        DEBUG_PRINTF("port_index=%u, universe=%u", port_index, universe);
+        DEBUG_PRINTF("port_index=%u, universe=%u", static_cast<unsigned>(port_index), static_cast<unsigned>(universe));
 
 #if defined(__GNUC__) && !defined(__clang__) && (__GNUC__ >= 13)
 #pragma GCC diagnostic pop
@@ -86,34 +78,28 @@ inline void SaveUniverse(uint32_t port_index)
 }
 } // namespace internal
 
-inline void SaveLongName(const char* long_name)
-{
+inline void SaveLongName(const char* long_name) {
     DEBUG_ENTRY();
     ConfigStore::Instance().DmxNodeUpdateArray(&common::store::DmxNode::node_name, long_name, artnet::kLongNameLength);
     DEBUG_EXIT();
 }
 
-inline void SaveShortName(uint32_t port_index, const char* short_name)
-{
+inline void SaveShortName(uint32_t port_index, const char* short_name) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("%u, %s", port_index, short_name);
+    DEBUG_PRINTF("%u, %s", static_cast<unsigned>(port_index), short_name);
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
@@ -123,34 +109,28 @@ inline void SaveShortName(uint32_t port_index, const char* short_name)
     DEBUG_EXIT();
 }
 
-inline void SaveSwitch(uint32_t port_index, [[maybe_unused]] uint8_t sw)
-{
+inline void SaveSwitch(uint32_t port_index, [[maybe_unused]] uint8_t sw) {
     DEBUG_ENTRY();
     internal::SaveUniverse(port_index);
     DEBUG_EXIT();
 }
 
-inline void SaveDirection(uint32_t port_index, dmxnode::Direction direction)
-{
+inline void SaveDirection(uint32_t port_index, dmxnode::Direction direction) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("%u, %u", port_index, static_cast<uint32_t>(direction));
+    DEBUG_PRINTF("%u, %u", static_cast<unsigned>(port_index), static_cast<unsigned>(direction));
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
@@ -164,27 +144,22 @@ inline void SaveDirection(uint32_t port_index, dmxnode::Direction direction)
     DEBUG_EXIT();
 }
 
-inline void SaveMergeMode(uint32_t port_index, dmxnode::MergeMode merge_mode)
-{
+inline void SaveMergeMode(uint32_t port_index, dmxnode::MergeMode merge_mode) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("%u, %u", port_index, static_cast<uint32_t>(merge_mode));
+    DEBUG_PRINTF("%u, %u", static_cast<unsigned>(port_index), static_cast<unsigned>(merge_mode));
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
@@ -198,27 +173,22 @@ inline void SaveMergeMode(uint32_t port_index, dmxnode::MergeMode merge_mode)
     DEBUG_EXIT();
 }
 
-inline void SaveProtocol(uint32_t port_index, artnet::PortProtocol port_protocol)
-{
+inline void SaveProtocol(uint32_t port_index, artnet::PortProtocol port_protocol) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("port_index=%u, portProtocol=%u", port_index, static_cast<uint32_t>(port_protocol));
+    DEBUG_PRINTF("port_index=%u, portProtocol=%u", static_cast<unsigned>(port_index), static_cast<unsigned>(port_protocol));
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
@@ -232,39 +202,31 @@ inline void SaveProtocol(uint32_t port_index, artnet::PortProtocol port_protocol
     DEBUG_EXIT();
 }
 
-inline void SaveOutputStyle(uint32_t port_index, dmxnode::OutputStyle output_style)
-{
+inline void SaveOutputStyle(uint32_t port_index, dmxnode::OutputStyle output_style) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("port_index=%u, output_style=%u", port_index, static_cast<uint32_t>(output_style));
+    DEBUG_PRINTF("port_index=%u, output_style=%u", static_cast<unsigned>(port_index), static_cast<unsigned>(output_style));
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
 
     auto output_style_store = ConfigStore::Instance().DmxNodeGet(&common::store::DmxNode::output_style);
 
-    if (output_style == dmxnode::OutputStyle::kConstant)
-    {
+    if (output_style == dmxnode::OutputStyle::kConstant) {
         output_style_store |= static_cast<uint8_t>(1U << port_index);
-    }
-    else
-    {
+    } else {
         output_style_store &= static_cast<uint8_t>(~(1U << port_index));
     }
 
@@ -273,27 +235,22 @@ inline void SaveOutputStyle(uint32_t port_index, dmxnode::OutputStyle output_sty
     DEBUG_EXIT();
 }
 
-inline void SaveRdmEnabled(uint32_t port_index, bool is_enabled)
-{
+inline void SaveRdmEnabled(uint32_t port_index, bool is_enabled) {
     DEBUG_ENTRY();
-    DEBUG_PRINTF("port_index=%u, is_enabled=%d", port_index, is_enabled);
+    DEBUG_PRINTF("port_index=%u, is_enabled=%u", static_cast<unsigned>(port_index), static_cast<unsigned>(is_enabled));
 
 #if defined(CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET)
-    {
+    if (port_index >= CONFIG_DMXNODE_DMX_PORT_OFFSET) {
         port_index -= CONFIG_DMXNODE_DMX_PORT_OFFSET;
-    }
-    else
-    {
+    } else {
         DEBUG_EXIT();
         return;
     }
 #endif
 
-    DEBUG_PRINTF("port_index=%u", port_index);
+    DEBUG_PRINTF("port_index=%u", static_cast<unsigned>(port_index));
 
-    if (port_index >= artnet::kPorts)
-    {
+    if (port_index >= artnet::kPorts) {
         DEBUG_EXIT();
         return;
     }
@@ -307,8 +264,7 @@ inline void SaveRdmEnabled(uint32_t port_index, bool is_enabled)
     DEBUG_EXIT();
 }
 
-inline void SaveFailSafe(uint8_t fail_safe)
-{
+inline void SaveFailSafe(uint8_t fail_safe) {
     DEBUG_ENTRY();
 
     ConfigStore::Instance().DmxNodeUpdate(&common::store::DmxNode::fail_safe, fail_safe);
